adding save all rubrique avec questions

+++ src/main/java/fr/ubo/dosi/projectagile/cscievaebackend/DTO/IncomingRubriqueQuestionDTO.java
@@ -7,6 +7,6 @@ import java.util.List;
 @Data
 public class IncomingRubriqueQuestionDTO {
     private Long idRubrique;
-    private List<Long> questionIds; // Assuming these are IDs of questions
+    private List<Integer> questionIds; // Assuming these are IDs of questions
     private Long ordre;
 }
